backend-drf/api/urls.py
from django.urls import path, include
from accounts import views as UserViews
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView 

urlpatterns = [
    path('register/', UserViews.RegisterView.as_view()), # as_view() is used to convert a class-based view into a view function that can be used in URL routing
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]