requirements.txt
django
python-decouple
djangorestframework
django-cors-headers
djangorestframework-simplejwt